Sticking point
stuck bec realized that output points, x,y of linkage when opt are not evenly distr. bec of the fact that the input is a velocity. Refer to iPad visualization. How to fix is to use Taylor approx to approx the curve, then have domain x with evenly distributed points. Or think of another way to do this.